Move deprecated hint processor
