+++ crates/blockifier/src/execution.rs
@@ -1,6 +1,5 @@
 pub mod common_hints;
 pub mod contract_class;
-pub mod deprecated_syscall_hint_processor;
 pub mod deprecated_syscalls;
 pub mod entry_point;
 pub mod errors;
